Actuator endpoint refactorings (no behavior changes) - Extract duplicated code into AddDiagnosticsManager() - Inline internal Add*ActuatorServices methods and reduce duplicated code into AddCoreActuatorServicesAsSingleton/Scoped - Split ActuatorServiceCollectionExtensions into CoreActuatorServiceCollectionExtensions and AllActuatorsServiceCollectionExtensions - Tweak the visibility of Core extension methods to enable reuse when building third-party actuators
