+++ src/Management/src/Abstractions/Diagnostics/ServiceCollectionExtensions.cs
@@ -1,0 +1,19 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information.
+
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+
+namespace Steeltoe.Management.Diagnostics;
+
+internal static class ServiceCollectionExtensions
+{
+    public static void AddDiagnosticsManager(this IServiceCollection services)
+    {
+        ArgumentNullException.ThrowIfNull(services);
+
+        services.TryAddSingleton<IDiagnosticsManager, DiagnosticsManager>();
+        services.AddHostedService<DiagnosticsService>();
+    }
+}
