+++ src/Management/src/Endpoint/Actuators/Metrics/EndpointServiceCollectionExtensions.cs
@@ -5,8 +5,11 @@
 using System.Diagnostics.Tracing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Steeltoe.Management.Diagnostics;
 using Steeltoe.Management.Endpoint.Actuators.Metrics.Observers;
+using Steeltoe.Management.Endpoint.Actuators.Metrics.SystemDiagnosticsMetrics;
 
 namespace Steeltoe.Management.Endpoint.Actuators.Metrics;
 
@@ -25,15 +28,60 @@ public static class EndpointServiceCollectionExtensions
     {
         ArgumentNullException.ThrowIfNull(services);
 
-        services.TryAddSingleton<IDiagnosticsManager, DiagnosticsManager>();
-        services.AddHostedService<DiagnosticsService>();
+        services.AddDiagnosticsManager();
 
-        services.AddCommonActuatorServices();
-        services.AddMetricsActuatorServices();
+        services
+            .AddCoreActuatorServicesAsSingleton<MetricsEndpointOptions, ConfigureMetricsEndpointOptions, MetricsEndpointMiddleware, IMetricsEndpointHandler,
+                MetricsEndpointHandler, MetricsRequest?, MetricsResponse?>();
 
+        services.TryAddSingleton<MetricsExporter>();
+
+        AddSteeltoeCollector(services);
         services.AddMetricsObservers();
 
         return services;
+    }
+
+    private static void AddSteeltoeCollector(IServiceCollection services)
+    {
+        services.AddSingleton(serviceProvider =>
+        {
+            var exporter = serviceProvider.GetRequiredService<MetricsExporter>();
+
+            MetricsEndpointOptions endpointOptions = serviceProvider.GetRequiredService<IOptionsMonitor<MetricsEndpointOptions>>().CurrentValue;
+            var logger = serviceProvider.GetRequiredService<ILogger<MetricsExporter>>();
+
+            var aggregationManager = new AggregationManager(endpointOptions.MaxTimeSeries, endpointOptions.MaxHistograms, exporter.AddMetrics,
+                (intervalStartTime, nextIntervalStartTime) => logger.LogTrace("Begin collection from {IntervalStartTime} to {NextIntervalStartTime}",
+                    intervalStartTime, nextIntervalStartTime),
+                (intervalStartTime, nextIntervalStartTime) => logger.LogTrace("End collection from {IntervalStartTime} to {NextIntervalStartTime}",
+                    intervalStartTime, nextIntervalStartTime),
+                instrument => logger.LogTrace("Begin measurements from {InstrumentName} for {MeterName}", instrument.Name, instrument.Meter.Name),
+                instrument => logger.LogTrace("End measurements from {InstrumentName} for {MeterName}", instrument.Name, instrument.Meter.Name),
+                instrument => logger.LogTrace("Instrument {InstrumentName} published for {MeterName}", instrument.Name, instrument.Meter.Name),
+                () => logger.LogTrace("Steeltoe metrics collector started."), exception => logger.LogError(exception, "An error occurred while collecting"),
+                () => logger.LogWarning("Cannot collect any more time series because the configured limit of {MaxTimeSeries} was reached",
+                    endpointOptions.MaxTimeSeries),
+                () => logger.LogWarning("Cannot collect any more Histograms because the configured limit of {MaxHistograms} was reached",
+                    endpointOptions.MaxHistograms), exception => logger.LogError(exception, "An error occurred while collecting observable instruments"));
+
+            exporter.SetCollect(aggregationManager.Collect);
+            aggregationManager.Include(SteeltoeMetrics.InstrumentationName); // Default to Steeltoe Metrics
+
+            foreach (string filter in endpointOptions.IncludedMetrics)
+            {
+                string[] filterParts = filter.Split(':');
+
+                if (filterParts.Length == 2)
+                {
+                    string meter = filterParts[0];
+                    string instrument = filterParts[1];
+                    aggregationManager.Include(meter, instrument);
+                }
+            }
+
+            return aggregationManager;
+        }).AddHostedService<MetricCollectionHostedService>();
     }
 
     /// <summary>
@@ -50,6 +98,7 @@ public static class EndpointServiceCollectionExtensions
         ArgumentNullException.ThrowIfNull(services);
 
         services.ConfigureOptionsWithChangeTokenSource<MetricsObserverOptions, ConfigureMetricsObserverOptions>();
+
         services.TryAddEnumerable(ServiceDescriptor.Singleton<IDiagnosticObserver, AspNetCoreHostingObserver>());
         services.TryAddEnumerable(ServiceDescriptor.Singleton<IDiagnosticObserver, HttpClientObserver>());
         services.TryAddEnumerable(ServiceDescriptor.Singleton<IRuntimeDiagnosticSource, ClrRuntimeObserver>());
